fix: packaging recent use cases
