+++ domain/src/main/java/com/london/domain/usecase/recent/search/ClearRecentSearchUseCase.kt
@@ -1,4 +1,4 @@
-package com.london.domain.usecase
+package com.london.domain.usecase.recent.search
 
 import com.london.domain.entity.recent.RecentSearch
 import com.london.domain.repository.RecentRepository
@@ -7,10 +7,10 @@ import org.koin.core.annotation.Provided
 import org.koin.core.annotation.Single
 
 @Single
-class DeleteRecentSearchUseCase(
+class ClearRecentSearchUseCase(
     @Provided
     @Named("recentSearchRepository")
     private val recentSearchRepository: RecentRepository<RecentSearch>
 ) {
-    suspend fun invoke(item: RecentSearch) = recentSearchRepository.delete(item)
+    suspend fun invoke() = recentSearchRepository.clearAll()
 }
