+++ domain/src/main/java/com/london/domain/usecase/recent/search/DeleteRecentSearchUseCase.kt
@@ -1,4 +1,4 @@
-package com.london.domain.usecase
+package com.london.domain.usecase.recent.search
 
 import com.london.domain.entity.recent.RecentSearch
 import com.london.domain.repository.RecentRepository
@@ -7,10 +7,10 @@ import org.koin.core.annotation.Provided
 import org.koin.core.annotation.Single
 
 @Single
-class ClearRecentSearchUseCase(
+class DeleteRecentSearchUseCase(
     @Provided
     @Named("recentSearchRepository")
     private val recentSearchRepository: RecentRepository<RecentSearch>
 ) {
-    suspend fun invoke() = recentSearchRepository.clearAll()
+    suspend fun invoke(item: RecentSearch) = recentSearchRepository.delete(item)
 }
