+++ domain/src/main/java/com/london/domain/usecase/recent/viewed/AddToRecentViewedUseCase.kt
@@ -1,4 +1,4 @@
-package com.london.domain.usecase
+package com.london.domain.usecase.recent.viewed
 
 import com.london.domain.entity.recent.RecentViewed
 import com.london.domain.repository.RecentRepository
@@ -7,10 +7,10 @@ import org.koin.core.annotation.Provided
 import org.koin.core.annotation.Single
 
 @Single
-class GetRecentViewedUseCase(
+class AddToRecentViewedUseCase(
     @Provided
     @Named("recentViewedRepository")
     private val repository: RecentRepository<RecentViewed>,
 ) {
-    suspend fun invoke() = repository.getAll()
+    suspend fun invoke(item: RecentViewed) = repository.insert(item)
 }
