+++ domain/src/main/java/com/london/domain/usecase/recent/watched/AddMovieToRecentWatchedUseCase.kt
@@ -1,16 +1,16 @@
-package com.london.domain.usecase
+package com.london.domain.usecase.recent.watched
 
-import com.london.domain.entity.TvShow
+import com.london.domain.entity.Movie
 import com.london.domain.repository.RecentWatchedRepository
 import org.koin.core.annotation.Named
 import org.koin.core.annotation.Provided
 import org.koin.core.annotation.Single
 
 @Single
-class AddTvShowToRecentWatchedUseCase(
+class AddMovieToRecentWatchedUseCase(
     @Provided
     @Named("recentWatchedRepository")
     private val recentWatchedRepository: RecentWatchedRepository
 ) {
-    suspend fun invoke(item: TvShow) = recentWatchedRepository.insertTvShow(item)
+    suspend fun invoke(item: Movie) = recentWatchedRepository.insertMovie(item)
 }
