+++ domain/src/test/kotlin/com/london/domain/usecase/GetRecentSearchUseCaseTest.kt
@@ -3,6 +3,7 @@ package com.london.domain.usecase
 import com.google.common.truth.Truth.assertThat
 import com.london.domain.entity.recent.RecentSearch
 import com.london.domain.repository.RecentRepository
+import com.london.domain.usecase.recent.search.GetRecentSearchUseCase
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
