+++ domain/src/test/kotlin/com/london/domain/usecase/GetRecentWatchedMoviesUseCaseTest.kt
@@ -3,6 +3,7 @@ package com.london.domain.usecase
 import com.google.common.truth.Truth.assertThat
 import com.london.domain.entity.Movie
 import com.london.domain.repository.RecentWatchedRepository
+import com.london.domain.usecase.recent.watched.GetRecentWatchedMoviesUseCase
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
