+++ presentation/src/main/java/com/london/presentation/screen/search/SearchViewModel.kt
@@ -5,16 +5,16 @@ import com.london.domain.entity.Movie
 import com.london.domain.entity.TvShow
 import com.london.domain.entity.recent.RecentSearch
 import com.london.domain.entity.recent.RecentViewed
-import com.london.domain.usecase.AddToRecentSearchUseCase
-import com.london.domain.usecase.AddToRecentViewedUseCase
-import com.london.domain.usecase.ClearRecentSearchUseCase
-import com.london.domain.usecase.ClearRecentViewedUseCase
-import com.london.domain.usecase.DeleteRecentSearchUseCase
+import com.london.domain.usecase.recent.search.AddToRecentSearchUseCase
+import com.london.domain.usecase.recent.viewed.AddToRecentViewedUseCase
+import com.london.domain.usecase.recent.search.ClearRecentSearchUseCase
+import com.london.domain.usecase.recent.viewed.ClearRecentViewedUseCase
+import com.london.domain.usecase.recent.search.DeleteRecentSearchUseCase
 import com.london.domain.usecase.GetActorsUseCase
 import com.london.domain.usecase.GetGenreInterestCountsUseCase
 import com.london.domain.usecase.GetMoviesUseCase
-import com.london.domain.usecase.GetRecentSearchUseCase
-import com.london.domain.usecase.GetRecentViewedUseCase
+import com.london.domain.usecase.recent.search.GetRecentSearchUseCase
+import com.london.domain.usecase.recent.viewed.GetRecentViewedUseCase
 import com.london.domain.usecase.GetTvShowsUseCase
 import com.london.domain.usecase.IncrementGenreInterestUseCase
 import com.london.presentation.screen.base.BaseViewModel
